HTML, JS and metadata updates to enable hiding 'request data' button
This enables hiding the 'request access' button via a dataset-level config option.
Ideally, the changes should be brought about by regenerating the metadata entries
and re-adding them to the catalog, with the new config as an added argument. Here,
changes are made directly to the metadata, which means a future cmmit will have to
do it the 'right way'.

The 'include_access_request' flag is introduced to the config file. If this is not
included for a dataset (i.e. all existing dataset metadata files) the pre-existing
logic is used for displaying the button, that is: if the 'access_request_contact'
or 'access_request_url' properties of the dataset have non-null values. If the
new flag included, its value is added as a logical AND to the pre-existing logic,
which then determines if the button is displayed or not.