Adds a script to generate a filelist for a dataset
This is mostly a copy of the code from 'dir2filetable' in datalad-tabby
The copied code has been updated to include a 'method' argument that will
tell the script which method to use when generating a file list. The
original method is called 'glob' and is also the default now. It runs on
a directory on the filesystem and uses standard Python libraries. The new
method is called 'tree' and uses datalad-next's 'tree' command. It is
therefore meant to be used for datalad datasets.